Refactor Globe and HomeHero for responsiveness and placeholder
